+++ src/main/java/net/sf/jabref/gui/ContentSelectorDialog2.java
@@ -27,7 +27,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
@@ -43,7 +42,6 @@ import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTextField;
 
-import net.sf.jabref.Globals;
 import net.sf.jabref.MetaData;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.gui.help.HelpFiles;
@@ -266,7 +264,7 @@ class ContentSelectorDialog2 extends JDialog {
         // If these were re-added, they will be added below, so it doesn't
         // cause any harm to remove them here.
         for (String fieldName : removedFields) {
-            metaData.remove(Globals.SELECTOR_META_PREFIX + fieldName);
+            metaData.clearContentSelectors(fieldName);
             changedFieldSet = true;
         }
 
@@ -284,22 +282,15 @@ class ContentSelectorDialog2 extends JDialog {
                     start++;
                 }
             }
-            List<String> data = metaData.getData(Globals.SELECTOR_META_PREFIX + fieldName);
-            boolean bNewField = false;
-            if (data == null) {
-                bNewField = true;
-                data = new ArrayList<>();
-                changedFieldSet = true;
 
-            } else {
-                data.clear();
-            }
-            for (int wrd = start; wrd < lm.size(); wrd++) {
-                String word = lm.get(wrd);
-                data.add(word);
-            }
-            if (bNewField) {
-                metaData.putData(Globals.SELECTOR_META_PREFIX + fieldName, data);
+            if (metaData.getContentSelectors(fieldName).isEmpty()) {
+                changedFieldSet = true;
+                List<String> data = new ArrayList<>();
+                for (int wrd = start; wrd < lm.size(); wrd++) {
+                    String word = lm.get(wrd);
+                    data.add(word);
+                }
+                metaData.setContentSelectors(fieldName, data);
             }
         }
 
@@ -324,8 +315,8 @@ class ContentSelectorDialog2 extends JDialog {
         fieldListModel.clear();
         SortedSet<String> contents = new TreeSet<>();
         for (String s : metaData) {
-            if (s.startsWith(Globals.SELECTOR_META_PREFIX)) {
-                contents.add(s.substring(Globals.SELECTOR_META_PREFIX.length()));
+            if (s.startsWith(MetaData.SELECTOR_META_PREFIX)) {
+                contents.add(s.substring(MetaData.SELECTOR_META_PREFIX.length()));
             }
         }
         if (contents.isEmpty()) {
@@ -363,14 +354,11 @@ class ContentSelectorDialog2 extends JDialog {
             wordListModel = new DefaultListModel<>();
             wordList.setModel(wordListModel);
             wordListModels.put(currentField, wordListModel);
-            List<String> items = metaData.getData(Globals.SELECTOR_META_PREFIX + currentField);
-            if (items != null) {
-                Set<String> wordSet = new TreeSet<>(items);
-                int index = 0;
-                for (String s : wordSet) {
-                    wordListModel.add(index, s);
-                    index++;
-                }
+
+            int index = 0;
+            for (String s : metaData.getContentSelectors(currentField)) {
+                wordListModel.add(index, s);
+                index++;
             }
         } else {
             wordList.setModel(wordListModel);